pyproject.toml
[tool.poetry]
name = "follow-my-reading"
version = "0.1.0"
description = "follow my reading: image, text, and audio processing"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "follow_my_reading"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.97.0"
pydantic = "^1.10.9"
uvicorn = "^0.22.0"
python-multipart = "^0.0.6"
pytesseract = "^0.3.10"
setuptools-rust = "^1.6.0"
openai-whisper = {git = "https://github.com/openai/whisper.git"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
black = "^23.3.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
